+++ packages/grafana-data/CHANGELOG.md
@@ -1,3 +1,9 @@
+# (Unreleased)
+
+### Bug fixes
+
+- Fixed crash when `LANG=c` is set in environment variables. Added locale validation to prevent crashes with invalid locales in DateTimeFormat. ([#110522](https://github.com/grafana/grafana/pull/110522))
+
 # (2019-07-08)
 
 First public release