+++ packages/grafana-i18n/CHANGELOG.md
@@ -1,3 +1,9 @@
+# (Unreleased)
+
+### Bug fixes
+
+- Fixed crash when `LANG=c` is set in environment variables. Added locale validation utilities to prevent crashes with invalid locales and provide safe fallbacks. ([#110522](https://github.com/grafana/grafana/pull/110522))
+
 # (2025-04-11)
 
 First public release